Fix warnings

+++ OHAttributedLabel/Source/OHTouchesGestureRecognizer.m
@@ -26,6 +26,7 @@
 
 
 #import "OHTouchesGestureRecognizer.h"
+#import "tgmath.h"
 
 #import <UIKit/UIGestureRecognizerSubclass.h>
 
@@ -49,7 +50,7 @@
     CGPoint currentPoint = [touch locationInView:self.view];
     CGFloat distanceX = (currentPoint.x - _startPoint.x);
     CGFloat distanceY = (currentPoint.y - _startPoint.y);
-    CGFloat distance = sqrtf(distanceX * distanceX + distanceY * distanceY);
+    CGFloat distance = sqrt(distanceX * distanceX + distanceY * distanceY);
     if (distance > 10.0f) {
         self.state = UIGestureRecognizerStateCancelled;
     } else {
